erts: Fix localtime BIFs does not update system timezone configuration
If in a running node the host machine changes its timezone
configuration, the BIFs time, date, localtime and
universal_time_to_localtime does not reflect the change until the node
is restarted.

The cause of this behaviour seems to be the lack of a
call to tzset() when the thread safe function is used.
The localtime manual page said:

   "According to POSIX.1-2004, localtime() is required to behave as
    though tzset() was called, while localtime_r() does not have
    this requirement.  For portable code tzset() should be called
    before localtime_r()."